requirements.txt
aiofiles==23.1.0
blinker==1.6.2
bottle==0.12.25
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
Flask==2.3.2
gevent==22.10.2
greenlet==2.0.2
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msvc-runtime==14.34.31931
multidict==6.0.4
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.1
pywebview==4.2.1
PyYAML==6.0
requests==2.31.0
sanic==23.3.0
sanic-ext==23.3.0
sanic-jinja2==2022.11.11
sanic-routing==22.8.0
tracerite==1.1.0
typing_extensions==4.6.3
urllib3==2.0.3
websocket==0.2.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0